Renaming tests. Add new directory with six files. Each file to navigate between pages.
