+++ locators.py
@@ -26,7 +26,7 @@ class Locator(object):
 
 
     #popup payments: EasyPay
-    
+    titlePopupEasyPay = "container > section > span:nth-child(3) > div > div > main > div > header > h2"
     mailInput = "div.StaggerGroup-child.is-head-0.is-tail-NaN > div > div > div > fieldset > span > div > div.Textbox-inputRow > input"
     cardNumberInput = "div.StaggerGroup-child.is-head-1.is-tail-NaN > span > span:nth-child(1) > div > div.Textbox-inputRow > input"
     dateCardInput = "div.Fieldset-childLeft.u-size1of2 > div.Textbox-inputRow > input"
